+++ CHANGELOG.md
@@ -8,4 +8,6 @@
 
 ### 🐛 Bug fixes
 
+- Register press-outside targets and blockers with `useNow` instead of render-time memo hooks.
+
 ### 💡 Others